+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Shortcode ```[breadcrumb]``` and ```[breadcrumb-item]```
 - Shortcode ```[button]``` and ```[button-group]``` and ```[button-toolbar]```
 - Shortcode ```[carousel]``` and ```[carousel-item]```
+- Shortcode ```[accordion]``` and ```[accordion-item]``` and ```[accordion-group]```